More reliable way to get TaskId
Previously the script would fail to create an AMI if there were multiple Import Volume tasks in the history, because it would retrieve the wrong VolumeId.
The proposed change introduces a function that uses awscli instead of ec2-api-tools to get the json information, passing through a python one liner to get the list of volume IDs.  This function is used in the loop where the status of import volume is checked before the script continues.