Extract vnd devices from vnconfig
The original code called vnconfig with just the image filename to load;
vnconfig was throwing an error (fresh install of OpenBSD6.5) asserting
that two parametes were required vnd_dev and image which were the vnd
device and image filename, respectively.  Therefore, to acquire the
vnd device, I used vnconfig -l to list the devices in-use, grepped
out the 'not in use' lines, grabbed the first one of them, and
extracted the vnd device from there.  As both instances of vnconfig
used _vndev to store the vnd device, I captured the output from the
aforementioned and passed it to vnconfig which seems to work.
Unfortunately, the installer panics later on when it's unable to
acquire the sets using https and the autoinstall response requires the
installer to not fallback to http.  That's.... another problem for
another day.

The error message is:

Unable to connect using https. Use http instead? [no] no